+++ minishell_sw/env/env_utils3.c
@@ -6,18 +6,18 @@
 /*   By: sunwsong <sunwsong@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/12 13:51:26 by sunwsong          #+#    #+#             */
-/*   Updated: 2023/02/14 17:39:21 by sunwsong         ###   ########.fr       */
+/*   Updated: 2023/02/15 20:21:26 by sunwsong         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-short	get_exitcode(void)
+int	get_exitcode(void)
 {
 	t_node	*cur;
 	int		exit_code;
 
 	cur = g_env_list->head->next;
-	exit_code = (short)ft_atoi(((t_env *)cur->val)->val);
+	exit_code = (int)ft_atoi(((t_env *)cur->val)->val);
 	return (exit_code);
 }
